Local variable table: tolerate flow from clashing variable defn
This can happen for the same reason as a previous local var table fix for flow from
out of range: jsr / ret control flow is imprecisely modelled, so control flow edges
are present which are in practice infeasible, hence the compiler generating code
which naively appears to change a stack slot's name or type without re-initialising it.

Consequently, demote these to warnings, and perhaps in the future do proper control-flow
analysis to determine which callsites can really reach which jsr subroutines.