Create an instruction for assigning the stub return to a variable
Previuosly we were returning the NONDET expression directly. However, we
were creating a symbol which was then not being used. This symbol's name
was then being used in the interpreter to work out what the function
should return.

This mirrors how the java test generation handles stubbing, but I think
we could probably do without this temporary variable as there is the
standard function_name#return_value variable.