Generalise dominators template
This previously assumed some internal specifics of goto_programt. This generalises it to allow the CFG graph to specify how to retrieve the start and end nodes or detect an empty program, and moves the details particular to goto_programt into its particular specialisation.

This allows us to use the generic dominators implementation on other control-flow graph representations.