Base-type-eq: check pointer types really match
Generally base-type-eq's task is to ensure that types match up to substitution of a type's body for a symbol referring to
that type. However now that derivatives of pointers exist with named subexpressions (e.g. java_generic_parametert), it is
now possible for types not involving symbols to be `base_type_eq` but not `==`. This resolves that particular case by
checking that the names subexpressions match exactly.