Add global name map to level2t
This global name map will be used to generate the next generation and shared across all states, while the current map that is copied to each state will be used for the local names.

When a query is done for the latest name, it will use the current names, so no matter if other parts of the code have generated new versions of this variable it will stay accurate to that branch. If you ask to increment the generation it'll update the global map and then set that name to the local one.

This means some vcc's may look a bit weird because a generation number may jump by quite a large margin (certainly in path symex) but this means that generation numbers don't need to be merged between branches, just work out if we need a phi node for them.