Refactor value_set_dereferencet
The aim is to make value_set_dereferencet::build_reference_to accessible
from the outside, by making it public and static. This required making
several other functions public and static as well.

The motivation is to use in value_set_element_to_expr, which is a helper
function for goto_symext::try_filter_value_sets.