Fix indirect Enum test
This never actually checked for the right thing; the test happened to pass because the return
value from Class.forName, when stubbed, might be null, and the actual code notionally under
test was always unreachable. Since the test is only here to ensure we don't have an invariant
violation when using an Enum which is never instantiated, I just remove the relevant condition
from the test.