CFG dominator algorithm change to more closely resemble Coopers
Main change is to make it so that cfg nodes only hold their immediate
dominators index, where previously it used to hold on to all dominators
up the tree. More compact (and faster) but if you need the full
dominator list then you have to walk back up dominator nodes.